change uint64 def

+++ types.hpp
@@ -3,7 +3,7 @@
 
 typedef unsigned short uint16;
 typedef unsigned int uint32;
-typedef unsigned long uint64;
+typedef unsigned long long uint64;
 
 static_assert(sizeof(uint16) == 2, "Change type of var for uint16");
 static_assert(sizeof(uint32) == 4, "Change type of var for uint32");
